Created Interpolation API
